Refactor traffic feature using inheritance
