+++ google_shop/models/__init__.py
@@ -23,4 +23,5 @@ from . import product_mapping
 from . import product
 from . import res_config_settings
 from . import target_country
-from . import product_traffic
+from . import google_shop_inheritance
+from . import product_traffic_inheritance
