+++ google_shop/models/product_traffic_inheritance.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from odoo import models, fields, api
+from odoo import models, fields
 
 class ProductTraffic(models.Model):
     _name = 'product.traffic'
