+++ google_shop/security/ir.model.access.csv
@@ -21,6 +21,4 @@ admin_field_property_data_manager,access_field_property_data,model_field_propert
 admin_property_field_mapping_line_manager,access_property_field_mapping_line,model_property_field_mapping_line,sales_team.group_sale_manager,1,1,1,1
 admin_property_data_name_manager,access_property_data_name,model_property_data_name,sales_team.group_sale_manager,1,1,1,1
 admin_google_shop_target_country,access_google_shop_target,model_target_country,sales_team.group_sale_manager,1,1,1,1
-user_product_traffic_user,product_traffic_user,model_product_traffic,google_shop.google_shop_user_group,1,1,1,1
-admin_product_traffic_manager,product_traffic_Manager,model_product_traffic,sales_team.group_sale_manager,1,1,1,1
