+++ google_shop/security/ir.model.access_inheritance.csv
@@ -1,0 +1,2 @@
+user_product_traffic_user,product_traffic_user,model_product_traffic,google_shop.google_shop_user_group,1,1,1,1
+admin_product_traffic_manager,product_traffic_Manager,model_product_traffic,sales_team.group_sale_manager,1,1,1,1
