+++ google_shop/views/product_traffic_view_inheritance.xml
